api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Oleksandr Pronin <pronin.alx@gmail.com>"]
readme = ""

[tool.poetry.dependencies]
python = "^3.11"
gql = {extras = ["all"], version = "^3.4.1"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.205.0"}
fastapi = "^0.103.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
